# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manu local HTTP API",
    "version": "v1",
    "description": "Contrato HTTP local, sem autenticação, para o primeiro recorte de ingestão, consulta e inspeção de evidências. A instância é limitada ao endereço loopback e usa uma única organização configurada; o valor padrão é `local`. Os envelopes JSON têm versão `v1`. Erros usam `application/problem+json` e não expõem conteúdo bruto, segredos ou diagnósticos de provedores."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Servidor local padrão (o endereço configurado deve permanecer em loopback)."
    }
  ],
...
  "paths": {
...
    "/api/v1/evidence/{evidence_id}": {
      "parameters": [{"$ref": "#/components/parameters/EvidenceID"}],
      "get": {
        "tags": ["evidence"],
        "summary": "Inspeciona uma Evidence Unit",
        "description": "Retorna somente identidade, proveniência, locator, estado de conteúdo e conteúdo local autorizado. Linhas omitidas ou negadas nunca carregam conteúdo. A coleção `/api/v1/evidence` não oferece operação de listagem neste recorte.",
        "operationId": "getEvidence",
        "responses": {
          "200": {
            "description": "Evidence Unit persistida e autorizada.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/EvidenceResponse"},
                "example": {
                  "version": "v1",
                  "id": "44444444-4444-4444-8444-444444444444",
                  "organization_id": "local",
                  "source_id": "22222222-2222-4222-8222-222222222222",
                  "snapshot_id": "33333333-3333-4333-8333-333333333333",
                  "artifact_id": "55555555-5555-4555-8555-555555555555",
                  "locator": {"path": "src/example.go", "start_line": 10, "end_line": 18},
                  "content_state": "present",
                  "classification": "safe_text",
                  "persist": "allow",
                  "external_transfer": "deny",
                  "content": "func Example() {}",
                  "content_hash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                  "content_bytes": 18,
                  "content_characters": 18,
                  "truncated": false,
                  "request_id": "req_example_20"
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidEvidencePath"},
          "404": {"$ref": "#/components/responses/EvidenceNotFound"},
          "408": {"$ref": "#/components/responses/RequestTimeout"},
          "500": {"$ref": "#/components/responses/EvidenceFailed"},
          "503": {"$ref": "#/components/responses/EvidenceUnavailable"},
          "405": {"$ref": "#/components/responses/MethodNotAllowedGet"}
        }
      }
    }
  },
  "components": {
    "headers": {
      "RequestID": {
        "description": "Identificador de correlação ecoado no corpo e na resposta. Se o request não o fornecer, o servidor gera um valor seguro.",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9_.-]{1,128}$"
        },
        "example": "req_example_01"
      }
    },
    "parameters": {
...
      "EvidenceID": {
        "name": "evidence_id",
        "in": "path",
        "required": true,
        "description": "Identificador UUID textual da Evidence Unit.",
        "schema": {"$ref": "#/components/schemas/UUID"},
        "example": "44444444-4444-4444-8444-444444444444"
      }
    },
    "responses": {
...
      "InvalidEvidencePath": {"$ref": "#/components/responses/Problem400EvidencePath"},
      "EvidenceNotFound": {"$ref": "#/components/responses/Problem404Evidence"},
      "EvidenceFailed": {"$ref": "#/components/responses/Problem500Evidence"},
      "EvidenceUnavailable": {"$ref": "#/components/responses/Problem503Evidence"},
      "RequestTimeout": {
        "description": "O request foi cancelado ou excedeu o prazo; `code` é `request_canceled` ou `request_timeout`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "request_timeout", "message": "request timed out", "status": 408, "request_id": "req_example_30"}}}
      },
...
      "MethodNotAllowedGet": {
        "description": "O método não é permitido para o recurso; `Allow: GET`.",
        "headers": {"Allow": {"schema": {"type": "string", "const": "GET"}}, "X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "method_not_allowed", "message": "method is not allowed", "status": 405, "request_id": "req_example_33"}}}
      },
...
      "Problem400EvidencePath": {
        "description": "O identificador ou caminho da evidência é inválido (`invalid_evidence_path`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "invalid_evidence_path", "message": "evidence path is invalid", "status": 400, "request_id": "req_example_48"}}}
      },
      "Problem404Evidence": {
        "description": "Evidence Unit não encontrada (`evidence_not_found`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "evidence_not_found", "message": "evidence was not found", "status": 404, "request_id": "req_example_49"}}}
      },
      "Problem500Evidence": {
        "description": "Evidence Unit armazenada inválida ou falha na inspeção (`evidence_invalid` ou `evidence_failed`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "evidence_failed", "message": "evidence inspection failed", "status": 500, "request_id": "req_example_50"}}}
      },
      "Problem503Evidence": {
        "description": "Serviço de evidências indisponível ou ocupado (`evidence_unavailable` ou `query_busy`); o limite global de concorrência pode usar `concurrency_limited`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "evidence_unavailable", "message": "evidence service is unavailable", "status": 503, "request_id": "req_example_51"}}}
      }
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
        "description": "UUID textual de 36 caracteres; a validação do handler aceita maiúsculas e minúsculas."
      },
      "Digest": {
        "type": "string",
        "pattern": "^[0-9a-f]{64}$",
        "description": "Digest SHA-256 em hexadecimal minúsculo."
      },
      "RequestID": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9_.-]{1,128}$"
      },
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "code", "message", "status", "request_id"],
        "properties": {
          "version": {"type": "string", "const": "v1"},
          "code": {"type": "string", "pattern": "^[a-z0-9_.-]{1,64}$"},
          "message": {"type": "string", "maxLength": 128},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "request_id": {"$ref": "#/components/schemas/RequestID"}
        },
        "description": "Envelope de erro seguro. O texto é fixo e não carrega conteúdo do request, DSN, segredo ou mensagem de provedor."
      },
...
      "Locator": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "uri": {"type": "string", "maxLength": 4096},
          "source_id": {"type": "string", "maxLength": 256},
          "artifact_id": {"type": "string", "maxLength": 256},
          "path": {"type": "string", "maxLength": 4096},
          "member": {"type": "string", "maxLength": 1024},
          "start_line": {"type": "integer", "minimum": 0},
          "start_column": {"type": "integer", "minimum": 0},
          "end_line": {"type": "integer", "minimum": 0},
          "end_column": {"type": "integer", "minimum": 0},
          "byte_offset": {"type": "integer", "minimum": 0},
          "byte_length": {"type": "integer", "minimum": 0}
        },
        "description": "Localizador de proveniência; não concede acesso ao filesystem original."
      },
      "EvidenceResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "id", "organization_id", "source_id", "snapshot_id", "artifact_id", "locator", "content_state", "classification", "persist", "external_transfer", "content_hash", "content_bytes", "content_characters", "truncated", "request_id"],
        "properties": {
          "version": {"type": "string", "const": "v1"},
          "id": {"$ref": "#/components/schemas/UUID"},
          "organization_id": {"type": "string", "minLength": 1},
          "source_id": {"$ref": "#/components/schemas/UUID"},
          "snapshot_id": {"$ref": "#/components/schemas/UUID"},
          "artifact_id": {"$ref": "#/components/schemas/UUID"},
          "observation_id": {"$ref": "#/components/schemas/UUID"},
          "locator": {"$ref": "#/components/schemas/Locator"},
          "content_state": {"type": "string", "enum": ["present", "redacted", "omitted"]},
          "classification": {"type": "string", "enum": ["", "safe_text", "sensitive", "prompt_injection_like", "binary", "invalid", "prohibited"]},
          "persist": {"type": "string", "enum": ["allow", "redact", "deny"]},
          "external_transfer": {"type": "string", "enum": ["allow", "redact", "deny"]},
          "content": {"type": "string", "maxLength": 65536, "description": "Presente apenas quando autorizado; nunca é serializado para conteúdo omitido ou negado."},
          "content_hash": {"$ref": "#/components/schemas/Digest"},
          "content_bytes": {"type": "integer", "minimum": 0},
          "content_characters": {"type": "integer", "minimum": 0},
          "truncated": {"type": "boolean"},
          "request_id": {"$ref": "#/components/schemas/RequestID"}
        },
        "description": "Proveniência e conteúdo local autorizado de uma Evidence Unit. O corpo nunca contém segredo, erro bruto ou conteúdo de uma unidade negada/omitida."
      }
    }
  }
}